tests/requirements/base.txt
pytest-django
pytest-cov
coverage[toml]
djangocms-versioning

